# Reduce allocs when setting value on `BindableNumber` with precision