BenchmarkDotNet isn't outputting a summary table, so I'll leave it up to @peppy to provide before/after tales. But the optimisation is pretty standard (see other cases of the same being done in this file).